% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiness_assessment_tests.R
\name{arrange_RAT_questions}
\alias{arrange_RAT_questions}
\title{Arrange RAT questions}
\usage{
arrange_RAT_questions(rat, by = "original", tRAT_n_offset = 0,
  fixed_number_column = NULL, group_by_column = NULL,
  random_seed = random())
}
\arguments{
\item{rat}{Readiness Assessment Test object}

\item{by}{what to arrange by, options: \code{by="original"} leaves the original order as encountered, \code{by="random"} generates a random order, \code{by="semi-random"} generates a random order except for questions that have the \code{fixed_number_column} set, and \code{by="fixed"} uses the \code{fixed_number_column} (which has to be set for all questions!).}

\item{tRAT_n_offset}{the offset for the question numbering (if there should be an offset to fit a different part of an IF-AT), only matters if \code{by} is not \code{"fixed"}. Default is 0 (so numbering starts at 1).}

\item{fixed_number_column}{name of a column in the RAT questions data frame that indicates the fixed question number. Only relevant if \code{by="semi-random"} or \code{by="fixed"}}

\item{group_by_column}{name of a column in the RAT questions data frame that indicates which questions to group together (groups will be arranged alphabetically, those with undefined group come last). Only relevant if \code{by="random"} or \code{by="semi-random"}. Throws an error if grouping and any fixed number questions are incompatible (e.g. a question has fixed number 5 but is part of the first group of only 3 questions).}

\item{random_seed}{can overwrite with a fixed value (e.g. \code{random_seed=42}) to get a reproducible "random" order in \code{by="random"} or \code{by="semi-random"} mode.}
}
\value{
returns the RAT with the questions having a new 'number' column
}
\description{
Arrange RAT questions
}
