% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_rows}
\alias{arrange_rows}
\title{Arrange rows}
\usage{
arrange_rows(.data, ..., desc = TRUE)
}
\arguments{
\item{.data}{data frame}

\item{...}{The name of up to two column names on which to arrange the rows}

\item{desc}{Logical indicating whether to arrange by descending (default) or
ascending values.}
}
\value{
Rearranged data frame
}
\description{
Arrange rows via descending or ascending column values
}
