% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{as_tbl}
\alias{as_tbl}
\title{as_tbl}
\usage{
as_tbl(x, row.names = FALSE)
}
\arguments{
\item{x}{Data}

\item{row.names}{Logical indicating whether to convert non-null row names
into the first column.}
}
\description{
See \code{tibble::\link[tibble]{as_tibble}} for details.
}
