% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_classroom.R
\name{authenticate_github}
\alias{authenticate_github}
\title{Authenticate GitHub account}
\usage{
authenticate_github(token)
}
\arguments{
\item{token}{authentication token for github API (https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/), requires scope "repo"}
}
\value{
GraphqlClient object
}
\description{
Authenticate GitHub account
}
