% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_cols.R
\name{bind_cols_data}
\alias{bind_cols_data}
\title{Bind columns}
\usage{
bind_cols_data(...)
}
\arguments{
\item{...}{Data frames}
}
\value{
Single data frame with combined columns
}
\description{
Bind together the columns of two or more data frames.
}
