% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_inventories.R
\name{bind_inventories}
\alias{bind_inventories}
\title{bind_inventories}
\usage{
bind_inventories(..., .id = "inventory", verbose = getOption("verbose"))
}
\arguments{
\item{...}{one or more tabular datasets}

\item{.id}{you can override this}

\item{verbose}{display messages}
}
\description{
Stack two inventories on top of one another.
}
