% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_classroom.R
\name{clone_classroom_repositories}
\alias{clone_classroom_repositories}
\title{Clone github classroom repositories}
\usage{
clone_classroom_repositories(token, org, repo_prefix, folder = ".",
  pull = TRUE)
}
\arguments{
\item{token}{authentication token for github API (https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/), requires scope "repo"}

\item{org}{organization name (i.e. organization github user)}

\item{repo_prefix}{repository prefix (i.e. prefix set of classroom repositories)}

\item{folder}{the target directory where to clone all the repositories to}

\item{pull}{if the repository already exists, whether to update it from the remote}
}
\value{
repositories data frame, see \link{find_classroom_repositories} for details
}
\description{
Clone github classroom repositories
}
