% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_classroom.R
\name{clone_repository}
\alias{clone_repository}
\title{Clone a repository to the target directory}
\usage{
clone_repository(url, path = NULL, token = NULL, pull = TRUE)
}
\arguments{
\item{url}{the https://... repository}

\item{path}{the directory to clone to, if not provided, clones into the current working directory with the repository name}

\item{token}{the authentication token (only required for private repositories)}

\item{pull}{if the repository already exists, whether to update it from the remote}
}
\description{
Clone a repository to the target directory
}
