% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiness_assessment_tests.R
\name{create_RAT_from_excel}
\alias{create_RAT_from_excel}
\title{Create Readiness Assessment Test from Excel
Load RAT questions and answer key from Excel. Note that logical columns are evaluated to TRUE if they are =1, =TRUE, ='TRUE', ='yes' or ='x', everything else is FALSE.}
\usage{
create_RAT_from_excel(filepath, questions_tab = "questions",
  key_tab = "key", filter_include = TRUE, fill_down_questions = TRUE)
}
\arguments{
\item{filepath}{the path to the excel file}

\item{questions_tab}{the name of the questions tab (requires at minimum columns 'question', 'answer', and logical TRUE/FALSE 'correct', plus logical 'include' if \code{filter_inclue=TRUE})}

\item{filter_include}{if set, only keeps questions that have the 'include' column set}

\item{fill_down_questions}{whether to fill down the questions column (i.e. if question an their parameters are only written in first row)}

\item{keys_tab}{the name of th keys tab (requires at mimimum columns 'number', 'option')}
}
\description{
Create Readiness Assessment Test from Excel
Load RAT questions and answer key from Excel. Note that logical columns are evaluated to TRUE if they are =1, =TRUE, ='TRUE', ='yes' or ='x', everything else is FALSE.
}
