% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decr.R
\name{decr}
\alias{decr}
\title{Decreasing order}
\usage{
decr(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
Reordered vector
}
\description{
Arranges vector in descending order
}
\examples{
## decreasing 1:10
decr(1:10)
}
