% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_distinct.R
\name{ensure_distinct}
\alias{ensure_distinct}
\title{Ensure uniqueness of records}
\usage{
ensure_distinct(input_data, ...)
}
\arguments{
\item{input_data}{data frame}

\item{\dots}{key vars (like \code{fac_id}, \code{src_id})}
}
\description{
Ensure uniqueness of records
}
