% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_join.R
\name{exact_join}
\alias{exact_join}
\title{Join two tbls together}
\usage{
exact_join(x, y, by = NULL, ..., fail_with = stop, convert = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x, }{y tbls to join}

\item{by}{a character vector of variables to join by}

\item{...}{see `help(join, package = "dplyr")`}

\item{convert}{passed to `dplyr::all_equal()`}

\item{verbose}{display messages}
}
\description{
This function will complain if x and y cannot be "exactly" joined.
This happens when there are either:

- rows in `x` with no match in `y`; or
- rows in `y` with no match in `x`
}
