% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lookup_table.R
\name{extract_lookup_table}
\alias{extract_lookup_table}
\title{Extract a lookup table from tabular data}
\usage{
extract_lookup_table(
  input_data,
  ...,
  na.rm = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{(tabular) containing 1+ "key vars" and exactly 1 "value var"}

\item{...}{(vars) \verb{key1, key2, key3, ..., value}}

\item{na.rm}{(logical)}

\item{verbose}{(logical)}
}
\description{
Extract a lookup table from tabular data
}
