% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree.R
\name{extract_tree}
\alias{extract_tree}
\title{Extract tree structure from a data.frame}
\usage{
extract_tree(input_data, form = "column_lineage", ...)
}
\description{
Extract tree structure from a data.frame
}
\note{
For use with "Maximum Entropy Summary Tree" R Markdown document(s)
}
