% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pollutants.R
\name{filter_pollutants}
\alias{filter_pollutants}
\title{filter_pollutants}
\usage{
filter_pollutants(
  input_data,
  pollutants,
  ...,
  .name = "pollutant",
  verbose = TRUE
)
}
\description{
Filter by pollutant, and (optionally) create custom labels at the same time.
}
\details{
Use with your own list (see examples below), or use with a curated list (like \code{DB_GHG_POLLUTANT_CODES}).
This can be particularly helpful when working with TACs (which may not have the \code{pol_abbr} that you want.)
}
\seealso{
\code{\link[=filter_categories]{filter_categories()}}

\code{\link[=filter_facilities]{filter_facilities()}}
}
