% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_rows}
\alias{filter_rows}
\title{Filter rows}
\usage{
filter_rows(.data, ...)
}
\arguments{
\item{.data}{Data frame or two dimensional array}

\item{...}{This should evaluate and reduce down to a numeric (row number)
or logical vector. Row numbers higher than what exists in x will be
ignored. Any numeric vector must be either all positive or all negative.}
}
\value{
Sliced/filtered data frame
}
\description{
Filter rows via integer/numeric position or logical vector
}
