% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_set.R
\name{filter_set}
\alias{filter_set}
\title{filter_set}
\usage{
filter_set(data, x, ...)
}
\arguments{
\item{data}{tabular data, like a \code{tibble} or \code{data.frame}}

\item{x}{set specification (see Details)}

\item{...}{reserved for future use}
}
\value{
a subset of \code{data} (possibly empty, or unchanged)
}
\description{
Filter tabular data using a set specification
}
\details{
\itemize{
\item \code{x} is presently expected to be tabular (e.g., a \code{data.frame} or \code{tibble}).
\item \code{x} and \code{data} must have at least column in common.
\item The result is essentially a \code{semi_join()} on the common columns.
\item If \code{x} is empty (zero rows) or missing, the result will just be \code{data}.
This is an attempt to behave analogously to `dplyr::filter()].
}
}
\seealso{
The BY2015 Conventions Paper doc.
}
