% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sources.R
\name{filter_sources}
\alias{filter_sources}
\title{Filter for specific sources at specific facilities}
\usage{
filter_sources(input_data, whitelist)
}
\arguments{
\item{input_data}{data frame containing column \code{fac_id}}

\item{whitelist}{named list, with facility names as names, and data frames as elements (see note and examples)}
}
\value{
subset of \code{input_data}, with additional column \code{fac_name}

subset of \code{input_data}, with additional columns \code{fac_name} and \code{src_name}
}
\description{
Filter for specific sources at specific facilities
}
