% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_years.R
\name{filter_years}
\alias{filter_years}
\title{filter_years}
\usage{
filter_years(input_data, years, verbose = getOption("verbose"))
}
\arguments{
\item{input_data}{tabular, with column \code{year}}

\item{years}{\code{RY}, \code{PY}, or \code{CY}}

\item{verbose}{display messages}
}
\description{
Placeholder for more sophisticated filtering (once \code{NPS} package is underway).
}
