% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_classroom.R
\name{find_classroom_repositories}
\alias{find_classroom_repositories}
\title{Find GitHub Classroom repositories}
\usage{
find_classroom_repositories(gqlc, org, repo_prefix, max = 100,
  quiet = FALSE)
}
\arguments{
\item{gqlc}{authenticated graphql client}

\item{org}{organization name (i.e. organization github user)}

\item{repo_prefix}{repository prefix (i.e. prefix set of classroom repositories)}

\item{max}{maximum number of repositories to retrieve}
}
\value{
data frame with repository names, urls, last commits, last pull requests
}
\description{
Find GitHub Classroom repositories
}
