% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{generate_access_codes}
\alias{generate_access_codes}
\title{function to generate random codes for students}
\usage{
generate_access_codes(n, length = 4)
}
\description{
function to generate random codes for students
}
