% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_by_data}
\alias{group_by_data}
\title{Group data}
\usage{
group_by_data(.data, ...)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Unquoted (non-standard evaluation) name(s) of group variable(s).}
}
\value{
A data frame with groups attribute
}
\description{
Indicate grouping variables in data frame
}
\examples{
d <- data.frame(a = c("a", "b", "c"), b = 1:3, stringsAsFactors = FALSE)
group_by_data(d, a)
}
