% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_data_data}
\alias{group_data_data}
\title{Group row numbers in grouped data}
\usage{
group_data_data(x)
}
\arguments{
\item{x}{Groupted data frame}
}
\value{
List of row numbers for each group
}
\description{
Returns row numbers for each group in grouped data
}
