% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_data_groups}
\alias{group_data_groups}
\title{Groups in grouped data}
\usage{
group_data_groups(x)
}
\arguments{
\item{x}{Grouped data frame}
}
\value{
Names of grouping variables
}
\description{
Returns grouping variable names of grouped data
}
