% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_tests_server.R
\name{immediate_feedback_test_server}
\alias{immediate_feedback_test_server}
\title{Immediate Feedback Test App Server}
\usage{
immediate_feedback_test_server(
  roster,
  questions,
  data_gs_title,
  gs_key_file,
  auto_login_access_code = NULL
)
}
\arguments{
\item{roster}{data frame with the roster and access codes}

\item{questions}{data frame with the questions}

\item{auto_login_access_code}{set an automatic login access code for testing and debugging purposes}
}
\description{
Generates the server part of the tRAT app
}
