% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_tests_ui.R
\name{immediate_feedback_test_ui}
\alias{immediate_feedback_test_ui}
\title{Immediate Feedback Test App UI}
\usage{
immediate_feedback_test_ui(app_title)
}
\arguments{
\item{app_title}{the title of the app, e.g. "Class name - tRAT #1"}
}
\description{
Generates the user interface part of the immediate feedback test app.
}
\keyword{internal}
