% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hierarchy.R
\name{import_hierarchy}
\alias{import_hierarchy}
\title{Import a hierarchy from a .CSV file}
\usage{
import_hierarchy(file, ..., verbose = TRUE)
}
\arguments{
\item{file, }{... (character) path to .CSV file}

\item{verbose}{(logical)}
}
\description{
Import a hierarchy from a .CSV file
}
\seealso{
\link{with_hierarchy} is better for general usage (and relies on this function).
}
