% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_if_data}
\alias{mutate_if_data}
\title{Mutate certain columns}
\usage{
mutate_if_data(.data, .predicate, .f)
}
\arguments{
\item{.data}{Input data frame}

\item{.predicate}{Function applied to each column evaluating to a logical}

\item{.f}{Function applied to each .predicate-passing column. Can be written
in the formula \code{~ .x} format.}
}
\value{
Data frame with .predicate-passing columns mutated.
}
\description{
Wrangle only columns that pass a logical test
}
\seealso{
Other mutate: \code{\link{mutate_data}}
}
\concept{mutate}
