% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_obs.R
\name{n_obs}
\alias{n_obs}
\title{Number of observersations}
\usage{
n_obs()
}
\value{
Number of observations in group
}
\description{
Returns number of observations in group (if no group then number of obs
in data frame)
}
