% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch}
\alias{patch}
\title{Patch data on the fly.}
\usage{
patch(object, cond, ...)
}
\arguments{
\item{object}{to be patched}

\item{cond}{logical condition(s) to be evaluated within scope of object}

\item{\dots}{name-value pairs}

\item{quiet}{suppress messages}
}
\description{
Patch data on the fly.
}
\note{
FIXME: support \code{where(...)} (again) in some way
}
