% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation_ui.R
\name{peer_evaluation_ui}
\alias{peer_evaluation_ui}
\title{Peer Evaluation App UI}
\usage{
peer_evaluation_ui(app_title)
}
\arguments{
\item{app_title}{the title of the Peer Evaluation App, e.g. "Class name - Peer Evaluation #1"}
}
\description{
Generates the user interface part of the peer evaluation app.
}
\keyword{internal}
