% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull_data}
\alias{pull_data}
\title{Pull data}
\usage{
pull_data(.data, ...)
}
\arguments{
\item{.data}{Input data frame}

\item{...}{Unquoted name of column to pull from data frame}
}
\value{
A column pulled from its data frame (inheriting whatever class the
  column is)
}
\description{
Pull (extract) data from a data frame
}
\examples{

## pull the 'y' variable from the data frame
tbl_data(x = rnorm(5), y = letters[1:5]) \%>\% pull_data(y)

}
