% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_distinct.R
\name{pull_distinct}
\alias{pull_distinct}
\title{pull_distinct}
\usage{
pull_distinct(x, ...)
}
\arguments{
\item{x}{tabular data}

\item{...}{arguments passed to \code{distinct()} and \code{pull()}}
}
\description{
\code{pull_distinct(x, var)} is equivalent to `pull(distinct(x, var), var).``
}
