% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_first.R
\name{pull_first}
\alias{pull_first}
\title{pull_first}
\usage{
pull_first(x, ...)
}
\arguments{
\item{x}{tibble}

\item{...}{passed to \code{\link[dplyr:pull]{dplyr::pull()}}}
}
\value{
object
}
\description{
pull_first
}
