% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_summary.R
\name{pull_summary}
\alias{pull_summary}
\title{pull_summary}
\usage{
pull_summary(x, ...)
}
\arguments{
\item{x}{tabular data}

\item{...}{arguments passed to \code{pull()}}
}
\description{
\code{pull_summary(x, var)} is equivalent to `summary(pull(x, var)).``
}
