% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_total.R
\name{pull_total}
\alias{pull_total}
\title{pull_total}
\usage{
pull_total(x, ...)
}
\arguments{
\item{x}{tabular data}

\item{...}{arguments passed to \code{pull()}}
}
\description{
\code{pull_total(x, var)} is equivalent to `total(pull(x, var)).``
}
