% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{Read from a CSV file}
\usage{
read_csv(file, ..., verbose = getOption("verbose"))
}
\arguments{
\item{file}{filename}

\item{...}{further arguments to \link[read_csv]{readr}}
}
\description{
Read from a CSV file
}
