% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ncdf.R
\name{read_ncdf}
\alias{read_ncdf}
\title{read_ncdf}
\usage{
read_ncdf(path, ..., verbose = getOption("verbose"))
}
\arguments{
\item{path}{character}

\item{...}{reserved for future use}

\item{verbose}{logical}
}
\description{
Import the contents of a NetCDF file as a \code{tibble}.
}
