% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xls.R
\name{read_xls}
\alias{read_xls}
\title{Read a worksheet from an XLS or XLSX file}
\usage{
read_xls(file, expect_md5 = NULL, sheet = 1, ...,
  verbose = getOption("verbose"))
}
\arguments{
\item{file}{filename}

\item{sheet}{name or number (starting from 1)}

\item{...}{further arguments to \link[read_excel]{readxl}}
}
\value{
\link[tbl_df]{dplyr}
}
\description{
Read a worksheet from an XLS or XLSX file
}
