% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_back}
\alias{repos_back}
\title{move vars to back}
\usage{
repos_back(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{columns to move to back}
}
\value{
Reordered data frame.
}
\description{
move vars to back
}
\examples{
## data with row names
d <- data.frame(x = rnorm(5), y = rnorm(5), row.names = letters[1:5])

## move x to back
repos_back(d, x)

}
