% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{repos_front}
\alias{repos_front}
\title{move vars to front}
\usage{
repos_front(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{columns to move to front}
}
\value{
Reordered data frame.
}
\description{
move vars to front
}
\examples{
## data with row names
d <- data.frame(x = rnorm(5), y = rnorm(5), row.names = letters[1:5])

## move y to front
repos_front(d, y)

}
