% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_lookup.R
\name{reverse_lookup}
\alias{reverse_lookup}
\title{Revalue a vector based on another (named) vector}
\usage{
reverse_lookup(x, table, ...)
}
\arguments{
\item{x}{vector}

\item{table}{named vector}
}
\description{
Revalue a vector based on another (named) vector
}
