% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_join.R
\name{rolling_join}
\alias{rolling_join}
\title{Emulate a rolling join}
\usage{
rolling_join(x, y, by, roll = Inf, allow.cartesian = TRUE, ...)
}
\arguments{
\item{x, y}{tbls to join}

\item{by}{a character vector of columns to join by. The last column will be rolled.}

\item{roll}{\code{Inf} to roll forward; \code{-Inf} to roll backwards.}

\item{\dots}{other parameters passed on to methods}
}
\description{
Emulate a rolling join
}
\note{
Uses \pkg{data.table} under the hood. See \code{\link[data.table]{[.data.table}} for details.
}
