% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{select_cols}
\alias{select_cols}
\title{Select columns}
\usage{
select_cols(.data, ...)
}
\arguments{
\item{.data}{Input data frame}

\item{...}{Unquoted names of columns to select}
}
\value{
Data frame with select columns
}
\description{
Select columns with non-standard evaluation
}
