% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_data}
\alias{select_data}
\title{Select columns}
\usage{
select_data(.data, ...)
}
\arguments{
\item{.data}{Input data frame}

\item{...}{Unquoted names of columns to select}
}
\value{
Data frame with select columns
}
\description{
Select columns with non-standard evaluation
}
