% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_last.R
\name{select_last}
\alias{select_last}
\title{Move selected column(s) to last position(s)}
\usage{
select_last(input_data, ...)
}
\arguments{
\item{input_data}{tabular data}

\item{...}{column name(s)}
}
\description{
Move selected column(s) to last position(s)
}
\seealso{
\code{\link[=select_first]{select_first()}}
}
