% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_data}
\alias{slice_data}
\title{Slice data}
\usage{
slice_data(.data, ...)
}
\arguments{
\item{.data}{Input data frame}

\item{...}{Expression to evaluate to integer row positions}
}
\value{
.data of evaluated row positions
}
\description{
Slice data by row number
}
\examples{

## data set
d <- tbl_data(x = rnorm(10), y = rnorm(10))

## slice first 4 rows
slice_data(d, 1:4)
}
