% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by.R
\name{sort_by_}
\alias{sort_by_}
\title{Sort by a summary (statistic) of a (grouping) variable}
\usage{
sort_by_(input_data, wt_var, group_var, wt_fun = median, na.rm = TRUE)
}
\description{
Sort by a summary (statistic) of a (grouping) variable
}
\examples{
abated_point_source_emissions(year = 2011:2014) \%>\%
 filter(pol_id == 41) \%>\% # Benzene
 annual_emissions_by(fac_id, pol_id) \%>\%
 sort_by_(wt_var = "ems_qty", group_var = "fac_id")

 @export
}
