% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_across.R
\name{sum_across}
\alias{sum_across}
\title{sum_across}
\usage{
sum_across(input_data, ..., na.rm = TRUE, .groups = "drop")
}
\arguments{
\item{input_data}{tabular data}

\item{...}{column name(s)}

\item{na.rm}{logical}

\item{.groups}{passed to \code{\link[dplyr:summarise]{dplyr::summarise()}}}
}
\description{
sum_across
}
