% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_data}
\alias{summarise_data}
\title{Summarise data}
\usage{
summarise_data(.data, ...)
}
\arguments{
\item{.data}{Data frame}

\item{...}{One or more namecd expressions evaluating with non-standard evaluation
to a single summary value.}
}
\value{
A summary-level data frame
}
\description{
Returns a summary-level data frame
}
