% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl_data}
\alias{tbl_data}
\title{tbl data}
\usage{
tbl_data(...)
}
\arguments{
\item{...}{A data frame, vector, or list of values of equal or single-value
length–similar to \link[base]{data.frame}.}
}
\value{
An object of class \code{c("tbl_data", "tbl_df", "tbl", "data.frame")}
}
\description{
Create a tibble data frame
}
