% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_frame.R
\name{tbl_data_frame}
\alias{tbl_data_frame}
\title{Create tbl_data_frame}
\usage{
tbl_data_frame(...)
}
\arguments{
\item{...}{Column vectors or expressions that reduce down to desired column vectors of
data frame}
}
\value{
A tbl_data data frame
}
\description{
Creates a data frame using non-standard evaluation
}
\examples{

## create data frame with two random variables
tbl_data_frame(
  a = rnorm(10),
  b = rnorm(10)
)

## create variables calculated using previous variables
tbl_data_frame(
  a = rnorm(10),
  b = rnorm(10),
  c = (a + b) / 2,
  d = a + b + c
)

}
