% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_tests.R
\name{tbl_deploy_immediate_feedback_test}
\alias{tbl_deploy_immediate_feedback_test}
\title{Deploy immediate feedback test app}
\usage{
tbl_deploy_immediate_feedback_test(
  folder = "immediate_feedback_test",
  appName = guess_from(folder),
  ...
)
}
\arguments{
\item{folder}{folder where the app is located (see \link{tbl_setup_immediate_feedback_test} for details)}

\item{appName}{name of the application for the web address on shiny.io. Can be provided manually but must be unique within an account. By default is guessed from the folder simply by removing special characters, making everything lower case, and replacing spaces with \code{_}.}

\item{...}{additional optional parameters passed on to \link[rsconnect]{deployApp}}
}
\description{
Upload the app to a shiny server via \link[rsconnect]{rsconnect}. This uses the rsconnect function \link[rsconnect]{deployApp} to upload or update a peer evaluation and thus requires rsconnect credentials to be already set. See \link[rsconnect]{setAccountInfo} or the \href{https://shiny.rstudio.com/articles/shinyapps.html#configure-rsconnect}{configuration help} for details on how to set your credentials for the \href{https://www.shinyapps.io}{shiny app platform}.
}
\seealso{
Other immediate feedback test functions: 
\code{\link{tbl_duplicate_immediate_feedback_test}()},
\code{\link{tbl_run_immediate_feedback_test}()},
\code{\link{tbl_setup_immediate_feedback_test}()},
\code{\link{tbl_test_immediate_feedback_test}()}
}
\concept{immediate feedback test functions}
