% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat.R
\name{tbl_deploy_team_rat}
\alias{tbl_deploy_team_rat}
\title{Deploy tRAT app}
\usage{
tbl_deploy_team_rat(folder = "tRAT", appName = guess_from(folder), ...)
}
\arguments{
\item{folder}{folder where the tRAT app is located (see \link{tbl_setup_team_rat} for details)}

\item{appName}{name of the application for the web address on shiny.io. Can be provided manually but must be unique within an account. By default is guessed from the folder simply by removing special characters, making everything lower case, and replacing spaces with \code{_}.}

\item{...}{additional optional parameters passed on to \link[rsconnect]{deployApp}}
}
\description{
Upload the tRAT app to a shiny server via \link[rsconnect]{rsconnect}. This uses the rsconnect function \link[rsconnect]{deployApp} to upload or update a peer evaluation and thus requires rsconnect credentials to be already set. See \link[rsconnect]{setAccountInfo} or the \href{https://shiny.rstudio.com/articles/shinyapps.html#configure-rsconnect}{configuration help} for details on how to set your credentials for the \href{https://www.shinyapps.io}{shiny app platform}.
}
\seealso{
Other tRAT functions: 
\code{\link{tbl_duplicate_team_rat}()},
\code{\link{tbl_run_team_rat}()},
\code{\link{tbl_setup_team_rat}()},
\code{\link{tbl_test_team_rat}()}
}
\concept{tRAT functions}
