% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_tests.R
\name{tbl_duplicate_immediate_feedback_test}
\alias{tbl_duplicate_immediate_feedback_test}
\title{Duplicate an immmediate feedback test app}
\usage{
tbl_duplicate_immediate_feedback_test(
  template = "immediate_feedback_test",
  folder = "immediate_feedback_test2",
  data_gs_title = NULL,
  overwrite = FALSE,
  check_gs_access = TRUE
)
}
\arguments{
\item{template}{folder where the existing app that is to be duplicated is located (path must be either relative to the current working directory or an absolute file path on the operating system).}

\item{folder}{target folder where to setup the duplicated app (path must be either relative to the current working directory or an absolute file path on the operating system). If the folder does not exist yet, it will be created automatically.}

\item{data_gs_title}{name of the google spreadsheet that should be used for storing the data for the immediate feedback test. This spreadsheet must already exist and the \code{gs_key_file} must grant access to it.}

\item{overwrite}{whether to overwrite the app in the target directory if it already exists}

\item{check_gs_access}{whether to confirm google spreadsheet access (using the \code{\link{tbl_check_gs_access}} function). Note that if this is set to \code{FALSE}, this function will NOT validate the \code{gs_key_file} and NOT check that the provided \code{data_gs_title} is a valid spreadsheet the key file grants access to.}
}
\value{
returns the \code{folder} invisibly for ease of use in pipelines
}
\description{
This functions makes it easy to duplicate an existing test by copying all relevant files into a new folder. You should change the google spreadsheet the new test is linked to by specifying the \code{data_gs_title} parameter (if you don't it will warn you but allow it). To create a new immediate feedback test from scratch, use \link{tbl_setup_immediate_feedback_test} instead.
}
\seealso{
Other immediate feedback test functions: 
\code{\link{tbl_deploy_immediate_feedback_test}()},
\code{\link{tbl_run_immediate_feedback_test}()},
\code{\link{tbl_setup_immediate_feedback_test}()},
\code{\link{tbl_test_immediate_feedback_test}()}
}
\concept{immediate feedback test functions}
