% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat.R
\name{tbl_duplicate_team_rat}
\alias{tbl_duplicate_team_rat}
\title{Duplicate a tRAT app}
\usage{
tbl_duplicate_team_rat(
  template = "tRAT",
  folder = "tRAT2",
  data_gs_title = NULL,
  overwrite = FALSE,
  check_gs_access = TRUE
)
}
\arguments{
\item{template}{folder where the existing tRAT that is to be duplicated is located (path must be either relative to the current working directory or an absolute file path on the operating system).}

\item{folder}{target folder where to setup the duplicated tRAT (path must be either relative to the current working directory or an absolute file path on the operating system). If the folder does not exist yet, it will be created automatically.}

\item{data_gs_title}{name of the google spreadsheet that should be used for storing the tRAT data. This spreadsheet must already exist and the \code{gs_key_file} must grant access to it.}

\item{overwrite}{whether to overwrite the app in the target directory if it already exists}

\item{check_gs_access}{whether to confirm google spreadsheet access (using the \code{\link{tbl_check_gs_access}} function). Note that if this is set to \code{FALSE}, this function will NOT validate the \code{gs_key_file} and NOT check that the provided \code{data_gs_title} is a valid spreadsheet the key file grants access to.}
}
\value{
returns the \code{folder} invisibly for ease of use in pipelines
}
\description{
This functions makes it easy to duplicate an existing tRAT by copying all relevant files into a new folder. You should change the google spreadsheet the new tRAT is linked to by specifying the \code{data_gs_title} parameter (if you don't it will warn you but allow it). To create a new tRAT from scratch, use \link{tbl_setup_team_rat} instead.
}
\seealso{
Other tRAT functions: 
\code{\link{tbl_deploy_team_rat}()},
\code{\link{tbl_run_team_rat}()},
\code{\link{tbl_setup_team_rat}()},
\code{\link{tbl_test_team_rat}()}
}
\concept{tRAT functions}
