% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_example_peer_evaluation}
\alias{tbl_example_peer_evaluation}
\alias{tbl_example_empty_peer_evaluation}
\alias{tbl_example_roster}
\title{Example files}
\usage{
tbl_example_peer_evaluation()

tbl_example_empty_peer_evaluation()

tbl_example_roster()
}
\description{
Example files
}
\details{
\code{tbl_example_peer_evaluation} returns the google spreadsheet key (\link{gs_key}) for an example peer evaluation dataset.

\code{tbl_example_empty_peer_evaluation} returns the google spreadsheet key (\link{gs_key}) for an example empty/newly set-up peer evaluation.

\code{tbl_example_roster} returns the example roster data frame
}
