% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat.R
\name{tbl_fetch_team_rat_data}
\alias{tbl_fetch_team_rat_data}
\title{Fetch the tRAT data}
\usage{
tbl_fetch_team_rat_data(
  folder = ".",
  data_gs_title = get_app_data_gs_title(folder),
  data_gs_id = NULL,
  teams = read_excel(file.path(folder, "teams.xlsx")),
  questions = suppressMessages(read_excel(file.path(folder, "questions.xlsx"))),
  gs_key_file = file.path(folder, "gs_key_file.json"),
  download_to = file.path(folder, "tRAT_data_downloaded.xlsx")
)
}
\arguments{
\item{folder}{folder where the tRAT app is located (relative to the location of the RMarkdown file if used in the latter context)}

\item{data_gs_title}{name of the google spreadsheet that should be used for storing the tRAT data. This spreadsheet must already exist and the \code{gs_key_file} must grant access to it.}

\item{data_gs_id}{optional alternative to the \code{data_gs_title}, a google spread sheet ID text or object (see \link[googlesheets4]{sheets_id}). If provided, it takes precedence over the \code{data_gs_title} parameter as the google spreadsheet doesn't need to be searched by name anymore (i.e. can be loaded faster).}

\item{teams}{data frame with the teams and access codes}

\item{questions}{data frame with the questions}

\item{gs_key_file}{path to your .json access key file for TBL google spreadsheets. See the \href{https://tbltools.kopflab.org/articles/peer_evaluations.html}{peer evaluations vignette} for details. This key file is safe to use on a secure shiny app server but be careful to never post this file anywhere publicly as it could be used to gain access to your TBL spreadsheets. Make sure to share the google spreadshhet for this tRAT with the \code{client_email} listed in the key file.}

\item{download_to}{location where the whole tRAT data sheet will be downloaded to for more efficient read access}
}
\description{
Fetches the tRAT data from the google spreadsheet and reads it (using \link{tbl_read_team_rat_data}). For standard installations of the tRAT app, all defaults should be sufficient.
}
