% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_generate_access_codes}
\alias{tbl_generate_access_codes}
\title{Generate random access codes}
\usage{
tbl_generate_access_codes(n, length = 4)
}
\arguments{
\item{n}{how many access codes to generate}

\item{length}{how many characters for each acccess code}
}
\value{
returns the access codes invisible
}
\description{
Utility function to generate a list of random access codes that can be copied into an Excel column.
}
