% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiness_assessment_tests.R
\name{tbl_generate_RAT_choices}
\alias{tbl_generate_RAT_choices}
\title{Generate RAT choices}
\usage{
tbl_generate_RAT_choices(rat, answer_layout = "vertical",
  answer_layout_column = NULL,
  answer_layout_funs = tbl_default_RAT_layouts(), random_seed = random())
}
\arguments{
\item{rat}{Readiness Assessment Test object}

\item{answer_layout}{how to arrange the answers, layouts supported by default are \code{"vertical"} and \code{"horizontal"} (recommended for image answers). The layout can be overwritten for individual questions by setting the \code{answer_layout_column} parameter. Custom answer layouts can be provided using the \code{answer_layout_funcs} parameter.}

\item{answer_layout_column}{set this parameter to a column name in the questions data frame that has a different layout name for questions that are indended to deviate from the default layout (\code{answer_layout}). All layouts must be defined in the \code{answer_layout_funs} (\code{"vertical"} and \code{"horizontal"} by default).}

\item{random_seed}{can overwrite with a fixed value (e.g. \code{random_seed=42}) to get a reproducible "random" order of the answer options}

\item{answer_layout_funcs}{Specify custom answer layouts by providing layout functions that differ from the default. See \code{tbl_default_RAT_layouts} for details on how these work.}
}
\description{
Generates the actual multiple choice questions in an RAT
}
