% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat.R
\name{tbl_generate_team_rat}
\alias{tbl_generate_team_rat}
\title{Generate tRATs}
\usage{
tbl_generate_team_rat(
  team_rat_data,
  correct = "#4DAF4A",
  incorrect = "#E41A1C",
  unknown = "#999999",
  width = 0.9,
  height = 0.9
)
}
\arguments{
\item{correct}{the color for correct answers (green by default)}

\item{incorrect}{the color for incorrect answers (red by default)}

\item{unknown}{the color for yet unknown answers (gray by default)}

\item{width}{the relative width of the option boxes (from 0 to 1)}

\item{height}{the relative height of the optin boxes (from 0 to 1)}

\item{data}{the tRAT data frame retrieved by \link{tbl_fetch_team_rat_data}}
}
\description{
This function generates visual tRATs based on the data retrieved by \link{tbl_fetch_team_rat_data} (the same function is used internally by the tRAT apps).
}
