% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_test.R
\name{tbl_read_immediate_feedback_test_data}
\alias{tbl_read_immediate_feedback_test_data}
\title{Read downloaded immediate feedback test data}
\usage{
tbl_read_immediate_feedback_test_data(
  folder = ".",
  roster = read_excel(file.path(folder, "roster.xlsx")),
  questions = suppressMessages(read_excel(file.path(folder, "questions.xlsx"))),
  download_file = file.path(folder, "immediate_feedback_test_data_downloaded.xlsx")
)
}
\arguments{
\item{folder}{folder where the immediate feedback test app is located (relative to the location of the RMarkdown file if used in the latter context)}

\item{roster}{data frame with the roster and access codes}

\item{questions}{data frame with the questions}

\item{download_file}{downloaded immediate feedback test data file, created by \link{tbl_fetch_immediate_feedback_test_data}}
}
\description{
Read immediate feedback test data that's already downloaded. Usually called indirectly via \link{tbl_fetch_immediate_feedback_test_data} unless there is no reason to fetch the latest data from the google spreadsheet.
}
