% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_read_peer_evaluation_data}
\alias{tbl_read_peer_evaluation_data}
\title{Read downloaded peer evaluation data}
\usage{
tbl_read_peer_evaluation_data(
  folder = ".",
  roster = read_excel(file.path(folder, "roster.xlsx")),
  download_file = file.path(folder, "pe_data_downloaded.xlsx")
)
}
\arguments{
\item{folder}{folder where the peer evaluation app is located (relative to the location of the RMarkdown file if used in the latter context)}

\item{roster}{data frame with the student roster}

\item{download_file}{downloaded peer evaluation data file, created by \link{tbl_fetch_peer_evaluation_data}}
}
\description{
Read peer evaluation data that's already downloaded. Usually called indirectly via \link{tbl_fetch_peer_evaluation_data} unless there is no reason to fetch the latest data from the google spreadsheet.
}
