% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat.R
\name{tbl_read_team_rat_data}
\alias{tbl_read_team_rat_data}
\title{Read downloaded tRAT data}
\usage{
tbl_read_team_rat_data(
  folder = ".",
  teams = read_excel(file.path(folder, "teams.xlsx")),
  questions = suppressMessages(read_excel(file.path(folder, "questions.xlsx"))),
  download_file = file.path(folder, "tRAT_data_downloaded.xlsx")
)
}
\arguments{
\item{folder}{folder where the tRAT app is located (relative to the location of the RMarkdown file if used in the latter context)}

\item{teams}{data frame with the teams and access codes}

\item{questions}{data frame with the questions}

\item{download_file}{downloaded tRAT data file, created by \link{tbl_fetch_team_rat_data}}
}
\description{
Read tRAT data that's already downloaded. Usually called indirectly via \link{tbl_fetch_team_rat_data} unless there is no reason to fetch the latest data from the google spreadsheet.
}
