% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_test.R
\name{tbl_run_immediate_feedback_test}
\alias{tbl_run_immediate_feedback_test}
\title{Run an immediate feedback test app}
\usage{
tbl_run_immediate_feedback_test(
  data_gs_title,
  roster = read_excel("roster.xlsx"),
  questions = suppressMessages(read_excel("questions.xlsx")),
  app_title = "immediate_feedback_test",
  auto_login_access_code = NULL,
  ...,
  launch = FALSE
)
}
\arguments{
\item{data_gs_title}{name of the google spreadsheet that should be used for storing the data for the immediate feedback test. This spreadsheet must already exist and the \code{gs_key_file} must grant access to it.}

\item{roster}{data frame with the roster and access codes}

\item{questions}{data frame with the questions}

\item{app_title}{the title of the app, e.g. "Class name - tRAT #1"}

\item{auto_login_access_code}{set an automatic login access code for testing and debugging purposes}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call (only if \code{launch = TRUE}), can include server-specific parameters such as host or port}

\item{launch}{whether to launch the app (TRUE) or return a shiny app object (FALSE) that then can be launched via \code{\link[shiny]{runApp}}}
}
\description{
This function starts the user interface. Note that this function is typically NOT called directly but indirectly by setting up the app using \link{tbl_setup_immediate_feedback_test}, adjusting the files in the app folder, and test running the app using \link{tbl_test_immediate_feedback_test}.
}
\seealso{
Other immediate feedback test functions: 
\code{\link{tbl_deploy_immediate_feedback_test}()},
\code{\link{tbl_duplicate_immediate_feedback_test}()},
\code{\link{tbl_setup_immediate_feedback_test}()},
\code{\link{tbl_test_immediate_feedback_test}()}
}
\concept{immediate feedback test functions}
