% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_run_peer_evaluation}
\alias{tbl_run_peer_evaluation}
\title{Run a peer evaluation app}
\usage{
tbl_run_peer_evaluation(data_gs_title,
  roster = read_excel("roster.xlsx"), gs_token = "gs_token.rds",
  app_title = "Peer Evaluation", welcome_md_file = "app_welcome.md",
  self_eval_plus_md_file = "app_self_evaluation_plus.md",
  self_eval_minus_md_file = "app_self_evaluation_minus.md",
  teammate_eval_plus_md_file = "app_teammate_evaluation_plus.md",
  teammate_eval_minus_md_file = "app_teammate_evaluation_minus.md",
  quant_scores_md_file = "app_quantitative_scores.md",
  points_per_teammate = 10, max_points = 15, min_points = 0,
  min_point_difference = 2, auto_login_access_code = NULL, ...,
  launch = FALSE)
}
\arguments{
\item{data_gs_title}{name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the credentials used when asked by this function must have write access to the spreadsheet.}

\item{roster}{data frame with the student roster}

\item{gs_token}{path to a google spreadsheet oauth 2.0 authentication token file (see \link[httr]{Token-class}). If none is provided or the file does not exist yet, will ask for google drive credentials interactively to generate a token for the peer evaluation app. The token is safe to use on a secure shiny app server but be careful to never post this token file anywhere publicly as it could be used to gain access to your google drive.}

\item{app_title}{the title of the Peer Evaluation App, e.g. "Class name - Peer Evaluation #1"}

\item{welcome_md_file}{path to a markdown (.md) file for the login welcome message}

\item{self_eval_plus_md_file}{markdown file for the "plus" self evaluation message}

\item{self_eval_minus_md_file}{markdown file for the "minus" self evaluation message}

\item{teammate_eval_plus_md_file}{markdown file for the "plus" teammate evaluation message}

\item{teammate_eval_minus_md_file}{markdown file for the "minus" teammate evaluation message}

\item{quant_scores_md_file}{markdown file for the quantiative scores message}

\item{points_per_teammate}{points per teammate}

\item{max_points}{the maximum number of points allowed per team member}

\item{min_points}{the smallest number of points allowed per team member}

\item{min_point_difference}{the minimum point difference required for the scores (set to 0 to allow all scores to be identical)}

\item{auto_login_access_code}{set an automatic login access code for testing and debugging purposes}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call (only if \code{launch = TRUE}), can include server-specific parameters such as host or port}

\item{launch}{whether to launch the app (TRUE) or return a shiny app object (FALSE) that then can be launched via \code{\link[shiny]{runApp}}}
}
\description{
This function starts the peer evaluation user interface. Note that this function is typically NOT called directly but indirectly by setting up the peer evalution app using \link{tbl_setup_peer_evaluation}, adjusting the files in the peer evaluation app folder, and test running the app using \link{tbl_test_peer_evaluation}.
}
\seealso{
Other peer evaluation functions: \code{\link{tbl_setup_peer_evaluation}},
  \code{\link{tbl_test_peer_evaluation}}
}
\concept{peer evaluation functions}
