% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat.R
\name{tbl_run_team_rat}
\alias{tbl_run_team_rat}
\title{Run a tRAT app}
\usage{
tbl_run_team_rat(
  data_gs_title,
  teams = read_excel("teams.xlsx"),
  questions = suppressMessages(read_excel("questions.xlsx")),
  app_title = "tRAT",
  auto_login_access_code = NULL,
  ...,
  launch = FALSE
)
}
\arguments{
\item{data_gs_title}{name of the google spreadsheet that should be used for storing the tRAT data. This spreadsheet must already exist and the \code{gs_key_file} must grant access to it.}

\item{teams}{data frame with the teams and access codes}

\item{questions}{data frame with the questions}

\item{app_title}{the title of the team RAT App, e.g. "Class name - tRAT #1"}

\item{auto_login_access_code}{set an automatic login access code for testing and debugging purposes}

\item{...}{passed on to the \code{\link[shiny]{runApp}} call (only if \code{launch = TRUE}), can include server-specific parameters such as host or port}

\item{launch}{whether to launch the app (TRUE) or return a shiny app object (FALSE) that then can be launched via \code{\link[shiny]{runApp}}}
}
\description{
This function starts the tRAT user interface. Note that this function is typically NOT called directly but indirectly by setting up the tRAT app using \link{tbl_setup_team_rat}, adjusting the files in the tRAT app folder, and test running the app using \link{tbl_test_team_rat}.
}
\seealso{
Other tRAT functions: 
\code{\link{tbl_deploy_team_rat}()},
\code{\link{tbl_duplicate_team_rat}()},
\code{\link{tbl_setup_team_rat}()},
\code{\link{tbl_test_team_rat}()}
}
\concept{tRAT functions}
