% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_test.R
\name{tbl_setup_immediate_feedback_test}
\alias{tbl_setup_immediate_feedback_test}
\title{Set up an immediate feedback test app}
\usage{
tbl_setup_immediate_feedback_test(
  folder = "immediate_feedback_test",
  data_gs_title = "immediate_feedback_test",
  template_roster_file = system.file(package = "tbltools", "extdata",
    "teams_template.xlsx"),
  template_questions_file = system.file(package = "tbltools", "extdata",
    "questions_template.xlsx"),
  gs_key_file = NULL,
  overwrite = FALSE,
  check_gs_access = TRUE
)
}
\arguments{
\item{folder}{target folder where to setup the immediate feedback test (path must be either relative to the current working directory or an absolute file path on the operating system). If the folder does not exist yet, it will be created automatically.}

\item{data_gs_title}{name of the google spreadsheet that should be used for storing the data for the immediate feedback test. This spreadsheet must already exist and the \code{gs_key_file} must grant access to it.}

\item{template_roster_file}{path to an excel (.xslx) file that contains the roster information to use as template for the  app (can be edited later)  - will use the package template by default}

\item{template_questions_file}{path to an excel (.xslx) file that contains the questions/answer key information to use as template for the immediate feedback app (can be edited later)  - will use the package template by default}

\item{gs_key_file}{path to your .json access key file for TBL google spreadsheets. See the \href{https://tbltools.kopflab.org/articles/peer_evaluations.html}{peer evaluations vignette} for details. This key file is safe to use on a secure shiny app server but be careful to never post this file anywhere publicly as it could be used to gain access to your TBL spreadsheets. Make sure to share the google spreadshhet for this immediate feedback test with the \code{client_email} listed in the key file.}

\item{overwrite}{whether to overwrite the app in the target directory if it already exists}

\item{check_gs_access}{whether to confirm google spreadsheet access (using the \code{\link{tbl_check_gs_access}} function). Note that if this is set to \code{FALSE}, this function will NOT validate the \code{gs_key_file} and NOT check that the provided \code{data_gs_title} is a valid spreadsheet the key file grants access to.}
}
\value{
returns the \code{folder} invisibly for ease of use in pipelines
}
\description{
This functions makes it easy to setup an immediate feedback test app folder. Immediate feedback tests are typically used for team readiness assessment tests (tRATs) in team-based-learning (TBL) but can be used for any type of immediate feedback assessment. It creates the app script \code{app.R} and copies the \code{roster.xslx} (an Excel spreadsheet), \code{questions.xlsx} (an Excel spreadsheet) as well as the necessary google access key file for the \code{data_gs_title} google spreadsheet to the same directory. The provided \code{gs_key_file} must grant access to the google spreadsheet (\code{data_gs_title}). See the \href{https://tbltools.kopflab.org/articles/peer_evaluations.html}{peer evaluations vignette} for details on how to set up google credentials and generate a key file. Unless \code{check_gs_access = FALSE}, this function will call the \code{\link{tbl_check_gs_access}} function to confirm that the key file works and provides access to the requested google spreadsheet. All \code{app.R}, \code{roster.xlsx}, \code{questions.xlsx} can be edited and tested in the app directory before uploading the whole app to a shiny server. On the shiny server, the \code{roster.xlsx} and \code{questions.xlsx} will be read-only and all actual data will be stored in the google spreadsheet data file.
}
\details{
Use \link{tbl_test_immediate_feedback_test} with the same \code{folder} as parameter to test run the application locally on your computer. It is recommended to always test the application locally before uploading it to the shiny application server.
}
\seealso{
Other immediate feedback test functions: 
\code{\link{tbl_deploy_immediate_feedback_test}()},
\code{\link{tbl_duplicate_immediate_feedback_test}()},
\code{\link{tbl_run_immediate_feedback_test}()},
\code{\link{tbl_test_immediate_feedback_test}()}
}
\concept{immediate feedback test functions}
