% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_setup_peer_evaluation}
\alias{tbl_setup_peer_evaluation}
\title{Set up a peer evaluation}
\usage{
tbl_setup_peer_evaluation(folder = "peer_evaluation",
  roster_file = system.file(package = "tbltools", "extdata",
  "roster_template.xlsx"), data_gs_title = "Peer Evaluation",
  gs_token = NULL, overwrite = FALSE)
}
\arguments{
\item{folder}{target folder where to setup the peer evaluation (path must be either relative to the current working directory or an absolute file path on the operating system). If the folder does not exist yet, it will be created automatically.}

\item{roster_file}{path to an excel (.xslx) file that contains the student roster information - will use the package template by default}

\item{data_gs_title}{name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the credentials used when asked by this function must have write access to the spreadsheet.}

\item{gs_token}{a google spreadsheet oauth 2.0 authentication token (see \link[httr]{Token-class}). If none is provided (\code{g_token = NULL}, the default), will ask for google drive credentials interactively to generate a token for the peer evaluation app. The token is safe to use on a secure shiny app server but be careful to never post this token file anywhere publicly as it could be used to gain access to your google drive.}

\item{overwrite}{whether to overwrite the app in the target directory if it already exists}
}
\value{
returns the \code{folder} invisibly for ease of use in pipelines
}
\description{
This functions makes it easy to setup a peer evaluation app folder. It creates the app script \code{app.R} and copies the \code{roster_file} (an Excel spreadsheet) as well as the necessary google drive credentials for the \code{data_gsheet} to the same directory. It will ask for google drive credentials using \link[googlesheets]{gs_auth} and the entered credentials must have access to the \code{data_gsheet}. Both \code{app.R} and the \code{roster_file} can be edited and tested in the app directory before uploading the whole app to a shiny server. On the shiny server, the \code{roster_file} will be read only and all actual data will be stored in the google drive data file.
}
\details{
Use \link{tbl_test_peer_evaluation} with the same \code{folder} as parameter to test run the peer evaluation application locally on your computer.
}
\seealso{
Other peer evaluation functions: \code{\link{tbl_run_peer_evaluation}},
  \code{\link{tbl_test_peer_evaluation}}
}
