% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiness_assessment_tests.R
\name{tbl_setup_RAT_demo}
\alias{tbl_setup_RAT_demo}
\title{Setup a demo RAT}
\usage{
tbl_setup_RAT_demo(overwrite = FALSE)
}
\arguments{
\item{overwrite}{whether to overwrite the folder if it already exists (default is not to overwrite anything)}
}
\description{
Setup a demo RAT
}
