% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiness_assessment_tests.R
\name{tbl_setup_RAT_template}
\alias{tbl_setup_RAT_template}
\title{Setup a new RAT template}
\usage{
tbl_setup_RAT_template(module = "module 1", n_questions = 10,
  n_options_per_q = 5, overwrite = FALSE)
}
\arguments{
\item{module}{name of the module, this will be used as the default folder and file name for the RAT. Can be a folder path in which case only the sub-directory name will be used for file names.}

\item{n_questions}{number of questions to template the RAT with}

\item{n_options_per_q}{default number of options per question}
}
\description{
This function templates an RAT. All parameters are merely the defaults used for templating the new RAT and all can be modified as needed in the RAT excel spreadsheet and RMarkdown that this function creates.
}
