% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_test.R
\name{tbl_summarize_immediate_feedback_test_data}
\alias{tbl_summarize_immediate_feedback_test_data}
\title{Summarize immediate feedback test data}
\usage{
tbl_summarize_immediate_feedback_test_data(
  data,
  guess_points = c(1, 0.5, 0.25)
)
}
\arguments{
\item{data}{the immediate feedback test data frame retrieved by \link{tbl_fetch_immediate_feedback_test_data}}

\item{guess_points}{points for each successive guess. The default is 1 point for getting the right answer on the first guess, 0.5 points on the second guess, 0.25 point on the third and 0 points thereafter.}
}
\description{
Summarizes the immediate feedback test data.
}
