% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat.R
\name{tbl_summarize_team_rat_data}
\alias{tbl_summarize_team_rat_data}
\title{Summarize tRAT data}
\usage{
tbl_summarize_team_rat_data(data, guess_points = c(4, 2, 1))
}
\arguments{
\item{data}{the tRAT data frame retrieved by \link{tbl_fetch_team_rat_data}}

\item{guess_points}{points for each successive guess. The default is 4 points for getting the right answer on the first guess, 2 points on the second guess, 1 point on the third and 0 points thereafter.}
}
\description{
Summarizes the tRAT data.
}
