% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immediate_feedback_tests.R
\name{tbl_test_immediate_feedback_test}
\alias{tbl_test_immediate_feedback_test}
\title{Test run immediate feedback test app locally}
\usage{
tbl_test_immediate_feedback_test(folder = "immediate_feedback_test", ...)
}
\arguments{
\item{folder}{folder where the app is located (see \link{tbl_setup_immediate_feedback_test} for details)}

\item{...}{optional parameters passed to \link[shiny]{runApp}}
}
\description{
This function starts an immediate feedback test app previously set up using \link{tbl_setup_immediate_feedback_test}. This provides a local version of the app before deploying it to a shiny apps server (using \link{tbl_deploy_immediate_feedback_test}). Keep in mind that the test application writes data to the google spreadsheet and make sure to delete any test records from the google spreadsheet that may interfere with any users' test taking prior to providing the students with the link to the deployed tRAT app.
}
\seealso{
Other immediate feedback test functions: 
\code{\link{tbl_deploy_immediate_feedback_test}()},
\code{\link{tbl_duplicate_immediate_feedback_test}()},
\code{\link{tbl_run_immediate_feedback_test}()},
\code{\link{tbl_setup_immediate_feedback_test}()}
}
\concept{immediate feedback test functions}
