% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_test_peer_evaluation}
\alias{tbl_test_peer_evaluation}
\title{Test run peer evaluation locally}
\usage{
tbl_test_peer_evaluation(folder = "peer_evaluation", ...)
}
\arguments{
\item{folder}{folder where the peer evaluation app is located (see \link{tbl_setup_peer_evaluation} for details)}

\item{...}{parameters passed to \link[shiny]{runApp}}
}
\description{
This function starts a peer evaluation app set up using \link{tbl_setup_peer_evaluation}.
}
\seealso{
Other peer evaluation functions: \code{\link{tbl_run_peer_evaluation}},
  \code{\link{tbl_setup_peer_evaluation}}
}
