% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat.R
\name{tbl_test_team_rat}
\alias{tbl_test_team_rat}
\title{Test run tRAT app locally}
\usage{
tbl_test_team_rat(folder = "tRAT", ...)
}
\arguments{
\item{folder}{folder where the tRAT app is located (see \link{tbl_setup_team_rat} for details)}

\item{...}{optional parameters passed to \link[shiny]{runApp}}
}
\description{
This function starts a tRAT app previously set up using \link{tbl_setup_team_rat}. This provides a local version of the app before deploying it to a shiny apps server (using \link{tbl_deploy_team_rat}). Keep in mind that the test application writes data to the google spreadsheet and make sure to delete any test records from the google spreadsheet that may interfere with any teams' tRAT prior to providing the students with the link to the deployed tRAT app.
}
\seealso{
Other tRAT functions: 
\code{\link{tbl_deploy_team_rat}()},
\code{\link{tbl_duplicate_team_rat}()},
\code{\link{tbl_run_team_rat}()},
\code{\link{tbl_setup_team_rat}()}
}
\concept{tRAT functions}
