% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tbltools-package}
\alias{tbltools}
\alias{tbltools-package}
\title{tbltools: Tools for Team-Based-Learning}
\description{
Tools for using Team-Based-Learning (TBL) approaches in the classroom.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/tbltools}
  \item Report bugs at \url{https://github.com/kopflab/tbltools/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})

}
\keyword{internal}
