% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat_server.R
\name{team_rat_server}
\alias{team_rat_server}
\title{Team Rapid Assessment Test App Server}
\usage{
team_rat_server(
  teams,
  questions,
  data_gs_title,
  gs_key_file,
  auto_login_access_code = NULL
)
}
\arguments{
\item{teams}{data frame with the teams and access codes}

\item{questions}{data frame with the questions}

\item{auto_login_access_code}{set an automatic login access code for testing and debugging purposes}
}
\description{
Generates the server part of the tRAT app
}
