% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_rat_ui.R
\name{team_rat_ui}
\alias{team_rat_ui}
\title{Team Rapid Assessment Test App UI}
\usage{
team_rat_ui(app_title)
}
\arguments{
\item{app_title}{the title of the team RAT App, e.g. "Class name - tRAT #1"}
}
\description{
Generates the user interface part of the team RAT app.
}
