% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{tidyselector}
\alias{tidyselector}
\title{tidyselector}
\usage{
tidyselector(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{vars to select}
}
\value{
Data with selected columns.
}
\description{
Select columns using tidy eval
}
\examples{
## data with row names
d <- data.frame(x = rnorm(5), y = rnorm(5), row.names = letters[1:5])

## select only x
tidyselector(d, x)

}
