% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{ungroup_data}
\alias{ungroup_data}
\title{Ungroup data}
\usage{
ungroup_data(.data)
}
\arguments{
\item{.data}{Grouped data}
}
\value{
Data grame without groups attribute
}
\description{
Ungroups grouped data
}
