% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_ids.R
\name{unpack_ids}
\alias{unpack_ids}
\title{Unpack a vector of "packed IDs" (like facility IDs or category IDs)}
\usage{
unpack_ids(input_data, id_var = NULL, verbose = FALSE)
}
\description{
Unpack a vector of "packed IDs" (like facility IDs or category IDs)
}
\examples{
input_data <- data_frame(cat_ids = c("c(1, 3)", "c(1:4)"))
unpack_ids(input_data, id_var = "cat_ids")
unpack_ids(input_data)

}
