% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.data.frame.R
\name{update.data.frame}
\alias{update.data.frame}
\title{update.data.frame}
\usage{
\method{update}{data.frame}(input_data, .where, ..., verbose = getOption("verbose"))
}
\description{
Analogous to SQL UPDATE command
}
