% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.list.R
\name{update.list}
\alias{update.list}
\title{Update the value of items in a list (possibly adding new items).}
\usage{
\method{update}{list}(object, ...)
}
\arguments{
\item{object}{a list}

\item{...}{arguments of the form key=value}
}
\value{
a copy of the list with updated items
}
\description{
Update the value of items in a list (possibly adding new items).
}
