% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_hierarchy.R
\name{validate_hierarchy}
\alias{validate_hierarchy}
\title{validate_hierarchy}
\usage{
validate_hierarchy(input_data, id_var = NULL, ...,
  verbose = getOption("verbose"))
}
\description{
Validates an imported or existing hierarchy.
}
\note{
Invoked automatically by \code{\link[=import_hierarchy]{import_hierarchy()}}.
}
\seealso{
\itemize{
\item \code{\link[=import_hierarchy]{import_hierarchy()}}
\item \code{\link[=with_hierarchy]{with_hierarchy()}}
}
}
