% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcard_join.R
\name{wildcard_join}
\alias{wildcard_join}
\title{Join tables by a single column, allowing for a wildcard character}
\usage{
wildcard_join(x, y, by, wildcard = "X", keep_pattern = FALSE, ...)
}
\arguments{
\item{x, y}{tbls to join}

\item{by}{name of variable to join by (only one is currently supported)}

\item{wildcard}{a single character (default is \code{?})}

\item{\dots}{other parameters passed onto methods}
}
\description{
Join tables by a single column, allowing for a wildcard character
}
