% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_column_sums.R
\name{with_column_sums}
\alias{with_column_sums}
\title{with_column_totals}
\usage{
with_column_sums(
  input_data,
  na.rm = TRUE,
  fill = "",
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{(tabular data)}

\item{verbose}{(logical) display messages}
}
\description{
Display totals for numeric columns.
}
\examples{
BY2011::BY2011_annual_emission_data \%>\% sample_n(3) \%>\% with_row_totals()

}
