% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_comment.R
\name{with_comment}
\alias{with_comment}
\title{Set the comment attribute on an object. Suitable for chaining.}
\usage{
with_comment(x, ...)
}
\arguments{
\item{x}{any \R object}

\item{value}{a \code{character} vector, or \code{NULL}.}
}
\description{
Set the comment attribute on an object. Suitable for chaining.
}
