% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_hierarchy.R
\name{with_hierarchy}
\alias{with_hierarchy}
\title{Annotate tabular data with hierarchical headings}
\usage{
with_hierarchy(
  input_data,
  using = NULL,
  depth = NULL,
  ...,
  warn_missing = FALSE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{tabular data}

\item{using}{tabular data, OR the name of a \code{.CSV} file found in \verb{BY2015/Hierarchies/}.}

\item{depth}{(integer) use at most this many levels}

\item{warn_missing}{(logical) warn if any records in \code{input_data} do not have a match in \code{using}}

\item{verbose}{(logical)}
}
\description{
Join hierarchy to \code{input_data}, yielding additional columns ending in \code{h1}, \code{h2}, \code{h3}, etc.
}
\details{
\itemize{
\item There must exist key column(s) in \code{input_data} (e.g. \code{cat_id} or \code{pol_id}) matching the key column(s) in \code{using}.
\item If \code{using} is a filename/path, then \code{import_hierarchy()} will be used to import, tidy, reshape, and validate it.
}
}
\note{
For historical reasons, \code{with_hierarchy()} effectively defaults to \code{with_category_hierarchy(using = t1327)}.
}
\seealso{
\itemize{
\item \code{\link[inventory:with_category_hierarchy]{inventory::with_category_hierarchy()}}
\item \code{\link[inventory:with_pollutant_hierarchy]{inventory::with_pollutant_hierarchy()}}
\item \code{\link[inventory:with_SIC_hierarchy]{inventory::with_SIC_hierarchy()}} and the \href{https://www.osha.gov/pls/imis/sic_manual.html}{OSHA SIC system} (\href{https://www.osha.gov/pls/imis/sicsearch.html}{search})
\item \code{\link[inventory:with_SCC8_hierarchy]{inventory::with_SCC8_hierarchy()}}, \code{\link[=with_SCC8_hierarchy]{with_SCC8_hierarchy()}}, and the \href{https://ofmpub.epa.gov/sccwebservices/sccsearch/docs/SCC-IntroToSCCs.pdf}{EPA SCC system} (\href{https://ofmpub.epa.gov/sccwebservices/sccsearch/}{search})
}
}
