% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pivot_table.R
\name{write_pivot_table}
\alias{write_pivot_table}
\title{Save a pivot table to disk}
\usage{
write_pivot_table(
  object,
  path,
  overwrite = TRUE,
  selfcontained = TRUE,
  ...,
  verbose = getOption("verbose")
)
}
\arguments{
\item{object}{created with \code{pivot_table()}}

\item{path}{(character) path on disk}

\item{...}{further arguments to \code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}}

\item{verbose}{(logical)}
}
\description{
Save a pivot table to disk
}
