% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{as_tbl}
\alias{as_tbl}
\title{as_tbl}
\usage{
as_tbl(x, row_names = FALSE)
}
\arguments{
\item{x}{Data frame or data frame-like input.}

\item{row_names}{Logical indicating whether to convert non-null row names
into the first column.}
}
\description{
Converts data objects to tibbles.
}
\examples{
## data with row names
d <- data.frame(x = rnorm(5), y = rnorm(5), row.names = letters[1:5])

## convert to tibble
as_tbl(d)

## convert to tibble and create row_names variable
as_tbl(d, row_names = TRUE)

}
