% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_classroom.R
\name{clone_repositories}
\alias{clone_repositories}
\title{Clone repositories}
\usage{
clone_repositories(repos, folder = ".", token = NULL, pull = TRUE)
}
\arguments{
\item{repos}{data frame with, at minimum, columns 'repository' (name) and 'url'}

\item{folder}{the target directory where to clone all the repositories to}

\item{token}{the authentication token (only required for private repositories)}

\item{pull}{if the repository already exists, whether to update it from the remote}
}
\description{
Clone repositories
}
