% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid.R
\name{expand_grid}
\alias{expand_grid}
\title{expand_grid}
\usage{
expand_grid(..., stringsAsFactors = FALSE)
}
\description{
Like \code{expand.grid} but treats \code{stringsAsFactors} as FALSE.
}
\examples{
expand_grid(`Foo Bar` = 1:3, baz = "bap")

}
