% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_categories.R
\name{filter_categories}
\alias{filter_categories}
\title{filter_categories}
\usage{
filter_categories(
  input_data,
  ...,
  .name = "category",
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{(tabular data)}

\item{...}{either an (optionally named) vector or list of category IDs, \emph{or} a tabular dataset having a column \code{cat_id}.}

\item{.name}{if the dot-args (above) are named, then the names will be used to populate a column in the result. \code{.name} will be that column's name.}
}
\description{
Filter by category, and (optionally) create custom labels at the same time.
}
\examples{
# Here are some emissions we'd like to filter AND label.
DB_data <- RY(2011:2012) \%>\% point_source_emissions()

# Here's what we want to analyze. (Hint: look at `DB_POLLUTANT_NAMES` for name-id pairs.)
RESIDENTIAL_NG_COMBUSTION_CATEGORIES <- c(
  "Space Heating" = 283,
  "Water Heating" = 284,
  "Cooking" = 285)

# `filter_categories()` both filters the data, and assigns names to `category` (by default; use the `.name` argument to change it).
residential_NG_data <- BY2011::BY2011_annual_emission_data \%>\% filter_categories(RESIDENTIAL_NG_COMBUSTION_CATEGORIES)
residential_NG_data \%>\% tabulate_emissions_by(category, cat_id, year)

}
\seealso{
\code{\link[=filter_facilities]{filter_facilities()}}

\code{\link[=filter_pollutants]{filter_pollutants()}}
}
