% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_join.R
\name{fun_join}
\alias{fun_join}
\title{(Left) join with a custom comparator}
\usage{
fun_join(left, right, by = intersect(names(left), names(right)),
  fun = `==`)
}
\arguments{
\item{left}{data.frame}

\item{right}{data.frame}

\item{by}{names of columns to join by}

\item{fun}{custom comparator (see examples)}
}
\description{
(Left) join with a custom comparator
}
\examples{
my_df <- data.frame(cyl = c(4, 4, 6, 8), vs = c(0, 1, NA, NA), foo = c("A", "B", "C", "D"))
my_fun <- function (e1, e2) (e1 == e2) | is.na(e2)
fun_join(mtcars, my_df, fun = my_fun)

}
