% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiness_assessment_tests.R
\name{generate_RAT_choices}
\alias{generate_RAT_choices}
\title{Generate RAT multiple choice questions}
\usage{
generate_RAT_choices(rat)
}
\arguments{
\item{rat}{Readiness Assessment Test object}
}
\description{
Generate RAT multiple choice questions
}
