% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{mutate_data}
\alias{mutate_data}
\title{Mutate data}
\usage{
mutate_data(.data, ...)
}
\arguments{
\item{.data}{Data frame}

\item{...}{One or more expressions using non-standard evaluation designed
to return a column of values. These should be named variables. Any names
already found in the input data will override those input data frame columns}
}
\value{
Mutated data frame
}
\description{
Wrangle data by adding or transforming columns
}
