% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table.R
\name{pivot_table}
\alias{pivot_table}
\title{Create a pivot table}
\usage{
pivot_table(input_data, rows, columns, values, aggregator = "Sum",
  renderer = "Heatmap", skip_years, inclusions = list(), menuLimit = 6000,
  file = NULL, sorters = list(pol_abbr = c("PM", "PM2.5", "PM10", "TOG",
  "ROG", "NOx", "SO2", "CO", "CO2", "CH4", "N2O", "HFC+PFC", "NH3")), ...,
  verbose = getOption("verbose"))
}
\arguments{
\item{rows}{(character) required (can be a single value or a vector)}

\item{columns}{(character) looks for `year`}

\item{values}{(character) looks for `ems_qty` or `tput_qty`}

\item{aggregator}{(character) passed to [rpivotTable::rpivotTable()]}

\item{renderer}{(character) passed to [rpivotTable::rpivotTable()]}

\item{skip_years}{(optional) drop some years}

\item{inclusions}{(list) passed to [rpivotTable::rpivotTable()]}

\item{menuLimit}{(integer) passed to [rpivotTable::rpivotTable()]}

\item{sorters}{(list) expert use only}

\item{...}{further arguments to [rpivotTable::rpivotTable()]}

\item{verbose}{(logical)}
}
\description{
Create a pivot table that's tailored for tabular emission data.
}
\seealso{
- `inventory` [conventions][inventory::conventions] for tabular data
}
