% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table.R, R/pivot_chart.R
\name{pivot_table}
\alias{pivot_table}
\alias{pivot_chart}
\title{pivot_table() and pivot_chart()}
\usage{
pivot_table(input_data, rows, columns, values, aggregator = "Sum",
  renderer = "Heatmap", skip_years, inclusions = list(),
  menuLimit = 6000, file = NULL, sorters = list(pol_abbr = c("PM",
  "PM2.5", "PM10", "TOG", "ROG", "NOx", "SO2", "CO", "CO2", "CH4", "N2O",
  "HFC+PFC", "NH3")), ..., verbose = getOption("verbose"))

pivot_chart(..., renderer = "Area Chart",
  verbose = getOption("verbose"))
}
\arguments{
\item{rows}{(character) required (can be a single value or a vector)}

\item{columns}{(character) looks for \code{year}}

\item{values}{(character) looks for \code{ems_qty} or \code{tput_qty}}

\item{aggregator}{(character) passed to \link[rpivotTable:rpivotTable]{rpivotTable()}}

\item{renderer}{(character) passed to \link[rpivotTable:rpivotTable]{rpivotTable()}}

\item{skip_years}{(optional) drop some years}

\item{inclusions}{(list) passed to \code{\link[rpivotTable:rpivotTable]{rpivotTable::rpivotTable()}}}

\item{menuLimit}{(integer) passed to \code{\link[rpivotTable:rpivotTable]{rpivotTable::rpivotTable()}}}

\item{sorters}{(list) expert use only}

\item{...}{further arguments to \code{\link[rpivotTable:rpivotTable]{rpivotTable::rpivotTable()}}}

\item{verbose}{(logical)}
}
\description{
Create an interactive pivot table (or chart). Tailored for emission-inventory data.
}
\seealso{
\itemize{
\item \code{inventory} \href{https://paper.dropbox.com/doc/BY2015-Conventions--Anuj_YBX9ORYU6LdCnlzOo_TAg-lbtgRJluUoJdnflc2VyKx}{conventions} for tabular data
\item the \href{https://github.com/smartinsightsfromdata/rpivotTable/blob/master/README.md}{rpivotTable} R package
\item the Javascript \href{https://pivottable.js.org/examples/}{PivotTable.js} library
}
}
