% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_first.R
\name{select_first}
\alias{select_first}
\title{Move selected column(s) to first position(s)}
\usage{
select_first(input_data, ...)
}
\arguments{
\item{input_data}{tabular data}

\item{...}{column name(s)}
}
\value{

}
\description{
Move selected column(s) to first position(s)
}
\seealso{
\code{\link[=select_last]{select_last()}}
}
