% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{sort_by_}
\alias{sort_by_}
\title{Sort by a summary (statistic) of a (grouping) variable}
\usage{
sort_by_(input_data, wt_var, group_var, wt_fun = median, na.rm = TRUE)
}
\description{
@export
@importFrom dplyr group_by_ desc
@importForm stats median
}
