% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_across.R
\name{sum_across}
\alias{sum_across}
\title{sum_across}
\usage{
sum_across(x, ..., na.rm = TRUE, .groups = "drop")
}
\arguments{
\item{.groups}{}
}
\value{

}
\description{
sum_across
}
\note{
May not work for list-columns. See TODO inside function.
}
