% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_check_gs_access}
\alias{tbl_check_gs_access}
\title{Check google spreadsheet access}
\usage{
tbl_check_gs_access(folder = "peer_evaluation",
  data_gs_title = "Peer Evaluation", gs_token = file.path(folder,
  "gs_token.rds"), new_credentials = FALSE)
}
\arguments{
\item{folder}{target folder where to setup the peer evaluation (path must be either relative to the current working directory or an absolute file path on the operating system). If the folder does not exist yet, it will be created automatically.}

\item{data_gs_title}{name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the credentials used when asked by this function must have write access to the spreadsheet.}

\item{gs_token}{path to a google spreadsheet oauth 2.0 authentication token file (see \link[httr]{Token-class}). If none is provided or the file does not exist yet, will ask for google drive credentials interactively to generate a token for the peer evaluation app. The token is safe to use on a secure shiny app server but be careful to never post this token file anywhere publicly as it could be used to gain access to your google drive.}

\item{new_credentials}{if set to TRUE, will always prompt the user to establish new google drive credentials. If FALSE (the default), will only do so if the \code{gs_token} file does not exist yet.}
}
\value{
returns the retrieved google spreadsheet key invisibly
}
\description{
This function checks whether google spreadsheet credentials are established and there is access to the \code{data_gs_title} google sheet. If the \code{gs_token} file does not exist yet or \code{new_credentials = TRUE}, it will establish new google drive credentials using \link[googlesheets]{gs_auth}.
}
\seealso{
Other peer evaluation functions: \code{\link{tbl_duplicate_peer_evaluation}},
  \code{\link{tbl_run_peer_evaluation}},
  \code{\link{tbl_setup_peer_evaluation}},
  \code{\link{tbl_test_peer_evaluation}}
}
\concept{peer evaluation functions}
