% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_deploy_peer_evaluation}
\alias{tbl_deploy_peer_evaluation}
\title{Deploy peer evaluation app}
\usage{
tbl_deploy_peer_evaluation(folder = "peer_evaluation", ...)
}
\arguments{
\item{folder}{folder where the peer evaluation app is located (see \link{tbl_setup_peer_evaluation} for details)}

\item{...}{passed on to \link[rsconnect]{deployApp}}
}
\description{
Upload the peer evaluation app to a shiny server via rsconnect. This requires rsconnect credentials to be already set --> see \link[rsconnect]{setAccountInfo}
}
