% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_deploy_peer_evaluation}
\alias{tbl_deploy_peer_evaluation}
\title{Deploy peer evaluation app}
\usage{
tbl_deploy_peer_evaluation(folder = "peer_evaluation", ...)
}
\arguments{
\item{folder}{folder where the peer evaluation app is located (see \link{tbl_setup_peer_evaluation} for details)}

\item{...}{optional parameters passed on to \link[rsconnect]{deployApp}}
}
\description{
Upload the peer evaluation app to a shiny server via \link[rsconnect]{rsconnect}. This uses the rsconnect function \link[rsconnect]{deployApp} to upload or update a peer evaluation and thus requires rsconnect credentials to be already set. See \link[rsconnect]{setAccountInfo} or the \href{https://shiny.rstudio.com/articles/shinyapps.html#configure-rsconnect}{configuration help} for details on how to set your credentials for the \href{https://www.shinyapps.io}{shiny app platform}.
}
\concept{peer evaluation functionsde}
