% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_fetch_peer_evaluation_data}
\alias{tbl_fetch_peer_evaluation_data}
\title{Fetch the peer evaluation data}
\usage{
tbl_fetch_peer_evaluation_data(
  folder = ".",
  data_gs_title = NULL,
  data_gs_id = NULL,
  data_gs_key = deprecated(),
  roster = read_excel(file.path(folder, "roster.xlsx")),
  gs_key_file = file.path(folder, "gs_key_file.json"),
  download_to = file.path(folder, "pe_data_downloaded.xlsx")
)
}
\arguments{
\item{folder}{folder where the peer evaluation app is located (relative to the location of the RMarkdown file if used in the latter context)}

\item{data_gs_title}{name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the \code{gs_key_file} must grant access to it.}

\item{data_gs_id}{optional alternative to the \code{data_gs_title}, a google spread sheet ID text or object (see \link[googlesheets4]{sheets_id}). If provided, it takes precedence over the \code{data_gs_title} parameter as the google spreadsheet doesn't need to be searched by name anymore (i.e. can be loaded faster).}

\item{roster}{data frame with the student roster}

\item{gs_key_file}{path to your .json access key file for peer evaluation google spreadsheets. See the \href{https://tbltools.kopflab.org/articles/peer_evaluations.html}{peer evaluations vignette} for details. This key file is safe to use on a secure shiny app server but be careful to never post this file anywhere publicly as it could be used to gain access to your peer evaluation spreadsheets. Make sure to share the google spreadshhet for this peer evaluation with the \code{client_email} listed in the key file.}

\item{download_to}{location where the whole peer evaluation data sheet will be downloaded to for more efficient read access}
}
\description{
Fetches the peer evaluation data from the google spreadsheet and reads it (using \link{tbl_read_peer_evaluation_data}). For standard installations of the peer evaluation app, all defaults should be sufficient with only parameter \code{data_gs_title} requiring a specification.
}
