% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_evaluation.R
\name{tbl_test_peer_evaluation}
\alias{tbl_test_peer_evaluation}
\title{Test run peer evaluation locally}
\usage{
tbl_test_peer_evaluation(folder = "peer_evaluation", ...)
}
\arguments{
\item{folder}{folder where the peer evaluation app is located (see \link{tbl_setup_peer_evaluation} for details)}

\item{...}{optional parameters passed to \link[shiny]{runApp}}
}
\description{
This function starts a peer evaluation app previously set up using \link{tbl_setup_peer_evaluation}. This provides a local version of the app before deploying it to a shiny apps server (using \link{tbl_deploy_peer_evaluation}). Keep in mind that the test application writes data to the google spreadsheet and make sure to delete any test records from the google spreadsheet that may interfere with any students' peer evaluations prior to providing the students with the link to the deployed peer evaluation app.
}
\seealso{
Other peer evaluation functions: \code{\link{tbl_check_gs_access}},
  \code{\link{tbl_duplicate_peer_evaluation}},
  \code{\link{tbl_run_peer_evaluation}},
  \code{\link{tbl_setup_peer_evaluation}}
}
\concept{peer evaluation functions}
