% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbltools-package.R
\docType{package}
\name{tbltools-package}
\alias{tbltools}
\alias{tbltools-package}
\title{tbltools: Tools for Working with Tibbles}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Tools for working with tibbles, or data.frame-like objects 
    generated by the 'tibble' package <https://tibble.tidyverse.org/>. Functions 
    assist in converting objects to tibbles and creating frequency tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/tbltools}
  \item Report bugs at \url{https://github.com/mkearney/tbltools/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
\keyword{internal}
