% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.data.frame.R
\name{update.data.frame}
\alias{update.data.frame}
\title{update.data.frame}
\usage{
\method{update}{data.frame}(input_data, .where, ..., verbose = getOption("verbose"))
}
\description{
Analogous to SQL UPDATE command
}
\examples{
RY2015_refining_data <- RY(2015) \%>\% point_source_emissions() \%>\% filter_facilities(DST_REFINING_FACILITIES)
RY2015_Chevron_FCCU_PM <- where(year == RY(2015), fac_id == 10, src_id == 4285, pol_abbr == "PM")
this_data <- RY2015_refining_data \%>\% update(RY2015_Chevron_FCCU_PM, ems_qty = ems_qty * 2.0)
this_data \%>\% tabulate_emissions_by(fac_id, pol_abbr, digits = 0)

}
