% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv.R
\name{write_csv}
\alias{write_csv}
\title{Write to a CSV file}
\usage{
write_csv(
  input_data,
  file,
  prompt = FALSE,
  signif = Inf,
  digits = 15,
  date_format = "\%Y-\%m-\%d",
  na = "",
  eol = "\\r\\n",
  row.names = FALSE,
  quote = TRUE,
  ...,
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{data frame}

\item{file}{filename}

\item{prompt}{\code{TRUE} or character; if not \code{FALSE}, confirmation is requested}

\item{signif}{(optional) significant digits (default \code{Inf})}

\item{digits}{(optional) passed to \code{format} if \code{signif} is not specified}

\item{na}{how to represent \code{NA} ("" means blank)}

\item{...}{further arguments to \link{write.csv}}
}
\value{
\preformatted{      \code{input_data}, possibly modified. See Note for details.
}
}
\description{
Write to a CSV file
}
\note{
In contrast to \link{write.csv}, \code{row.names} is set to
\code{FALSE}, so that an extra column will not be written.

If \code{signif} is supplied, then the values that are written will be rounded.
It is these rounded values that are returned (invisibly, for chaining).
}
\seealso{
\link{read_tbl}
}
