% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pivot_table.R
\name{write_pivot_table}
\alias{write_pivot_table}
\title{Save a pivot table to disk}
\usage{
write_pivot_table(object, file, overwrite = TRUE, selfcontained = TRUE,
  ..., verbose = getOption("verbose"))
}
\arguments{
\item{object}{created with `pivot_table()`}

\item{file}{(character) path on disk}

\item{...}{further arguments to [htmlwidgets::saveWidget()]}

\item{verbose}{(logical)}
}
\description{
Save a pivot table to disk
}
