% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_file.R
\name{zip_file}
\alias{zip_file}
\title{Write to ZIP file}
\usage{
zip_file(x, file, flags = "-q", clean = TRUE, ...)
}
\arguments{
\item{x}{name(s) of input file(s)}

\item{file}{name of ZIP file (see note)}

\item{flags}{flags passed to \link{zip} ("q" means "quiet")}

\item{clean}{whether to remove input file(s) afterwards}

\item{...}{further arguments to \link{zip}}
}
\value{
\preformatted{      \code{file}, invisibly (for chaining)
}
}
\description{
Write to ZIP file
}
\seealso{
\link{write_tbl}
}
